Updated TaskCard.jsx & api/tasks/
Updated delete task function and buttona and adjusted the api/task/index.js file to handle the delete request. Task will now be deleted successfully with the deletetask function

+++ src/components/TaskCard.jsx
@@ -49,7 +49,14 @@ export function TaskCard(props) {
     };
     const handleDeleteTask = async () => {
         try {
-            const response = await fetch(`/api/tasks/${task._id}`, {
+            console.log(task._id);
+            console.log(task)
+            if (!task._id) {
+                console.error('Task does not have an id');
+                return;
+            }
+
+            const response = await fetch(`/api/tasks?id=${task._id}`, {
                 method: 'DELETE',
                 headers: {
                     'Content-Type': 'application/json',
